@@ -11,26 +11,25 @@
     },
     "license": "ISC",
     "dependencies": {
-        "@angular/common": "2.0.0-rc.1",
-        "@angular/compiler": "2.0.0-rc.1",
-        "@angular/core": "2.0.0-rc.1",
-        "@angular/http": "2.0.0-rc.1",
-        "@angular/platform-browser": "2.0.0-rc.1",
-        "@angular/platform-browser-dynamic": "2.0.0-rc.1",
-        "@angular/router": "2.0.0-rc.1",
-        "@angular/router-deprecated": "2.0.0-rc.1",
-        "@angular/upgrade": "2.0.0-rc.1",
-        "systemjs": "0.19.27",
-        "core-js": "^2.4.0",
+        "@angular/common": "2.0.0",
+        "@angular/compiler": "2.0.0",
+        "@angular/core": "2.0.0",
+        "@angular/forms": "2.0.0",
+        "@angular/http": "2.0.0",
+        "@angular/platform-browser": "2.0.0",
+        "@angular/platform-browser-dynamic": "2.0.0",
+        "@angular/router": "3.0.0",
+        "systemjs": "0.19.38",
+        "core-js": "^2.4.1",
         "reflect-metadata": "^0.1.3",
-        "rxjs": "5.0.0-beta.6",
-        "zone.js": "^0.6.12",
+        "rxjs": "5.0.0-beta.12",
+        "zone.js": "^0.6.23",
         "esri-system-js": "^1.0.0-beta.0"
     },
     "devDependencies": {
-        "concurrently": "^2.0.0",
-        "lite-server": "^2.2.0",
-        "typescript": "^1.8.10",
-        "typings": "^1.0.4"
+        "concurrently": "^2.2.0",
+        "lite-server": "^2.2.2",
+        "typescript": "^2.0.2",
+        "typings": "^1.3.2"
     }
 }
