@@ -13,7 +13,7 @@ import { Component } from '@angular/core';
                     <ul>
                         <li>MapView and SceneView components</li>
                         <li>Custom view coordination service to keep SceneView components in sync</li>
-                        <li>Vector GIS analysis in the browser</li>
+                        <li>Vector GIS analysis in the browser with reactive forms</li>
                     </ul>
                 </li>
                 <li>and <a href="http://picnicss.com/">Picnic CSS</a> for the good looks</li>
